+++ public/backend/App/Controller/ControllerAuth.php
@@ -5,43 +5,14 @@
     use App\Model\User;
     use App\DAO\UserDAO;
 
+    use Src\Classes\Authorization;
     use Src\Classes\Token;
 
     /**
-     * This class is responsible for business rules involving access validation. 
+     * This class is responsible for business rules involving access validation.
      */
     
     class ControllerAuth extends Controller{
-
-        public function getAuthorizationHeader(){
-            $headers = null;
-            
-            if (isset($_SERVER['Authorization'])) {
-                $headers = trim($_SERVER["Authorization"]);
-            }else if(isset($_SERVER['HTTP_AUTHORIZATION'])){
-                $headers = trim($_SERVER["HTTP_AUTHORIZATION"]);
-            }else if(function_exists('apache_request_headers')) {
-                $requestHeaders = apache_request_headers();
-                $requestHeaders = array_combine(array_map('ucwords', array_keys($requestHeaders)), array_values($requestHeaders));
-                // print_r($requestHeaders);
-                if (isset($requestHeaders['Authorization'])) {
-                    $headers = trim($requestHeaders['Authorization']);
-                }
-            }
-            return $headers;
-        }
-
-        private function getBearerToken() {
-            $headers = $this->getAuthorizationHeader();
-
-            // HEADER: Get the access token from the header
-            if (!empty($headers)) {
-                if (preg_match('/Bearer\s(\S+)/', $headers, $matches)) {
-                    return $matches[1];
-                }
-            }
-            return null;
-        }
         
         public function login($data){
 
@@ -69,14 +40,14 @@
         }
 
         public function validate(){
-            $token = $this->getBearerToken();
+            $token = (new Authorization())->getBearerToken();
             $data = (new Token())->get_token_data($token);
             
             return (new Token())->validate($token);
         }
 
         public function get_token_data(){
-            $token = $this->getBearerToken();
+            $token = (new Authorization())->getBearerToken();
             return (new Token())->get_token_data($token);
         }
     }
